apps/databridge-researcher/requirements.txt
# ============================================
# DataBridge Analytics - Python Dependencies
# ============================================
# Install: pip install -r requirements.txt
# ============================================

# ----------------
# CLI Framework
# ----------------
typer>=0.12.0                   # Modern CLI with type hints
click>=8.1.0                    # CLI utilities
rich>=13.7.0                    # Beautiful terminal output
prompt-toolkit>=3.0.0           # Interactive REPL

# ----------------
# Data Processing
# ----------------
pandas>=2.2.0                   # DataFrame operations
numpy>=1.26.0                   # Numerical operations
pyarrow>=15.0.0                 # Efficient data transfer

# ----------------
# Database Connectivity
# ----------------
sqlalchemy>=2.0.0               # Database abstraction
alembic>=1.13.0                 # Database migrations

# ----------------
# Configuration
# ----------------
pydantic>=2.6.0                 # Data validation
pydantic-settings>=2.2.0        # Settings management
python-dotenv>=1.0.0            # .env file loading

# ----------------
# MCP Server
# ----------------
fastmcp>=0.4.0                  # MCP framework for Claude

# ----------------
# Vector Embeddings & Semantic Search
# ----------------
chromadb>=0.4.22                # Vector database
sentence-transformers>=2.5.0    # Local embeddings

# ----------------
# HTTP Client
# ----------------
httpx>=0.27.0                   # Async HTTP client

# ----------------
# Utilities
# ----------------
python-slugify>=8.0.0           # Slug generation
arrow>=1.3.0                    # Date/time utilities
tabulate>=0.9.0                 # Table formatting
tqdm>=4.66.0                    # Progress bars

# ----------------
# Security
# ----------------
cryptography>=42.0.0            # Encryption for credentials

# ----------------
# PostgreSQL (Docker Sample Database)
# ----------------
psycopg2-binary>=2.9.9          # PostgreSQL driver
asyncpg>=0.29.0                 # Async PostgreSQL
